Gameplay ingestion for active games
This is probably the last script in python, passively compilates gameplay information for all the user's active game play registers. If the user sets profile to private is immediatly banned (B02: two months of ban) and if the user lies about having a game a strike is triggered and he/she will be kicked from all J01 and J02 class jackpots.